refactor: modify import route
