+++ src/components/features/score/list/form/YoutubeSearchWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 
 import { Input } from "@/components/ui/input";
-import useDebounce from "@/hooks/useDebounce/useDebounce";
+import useDebounce from "@/hooks/useDebounce";
 import { YoutubeSearchItem } from "@/type/youtube";
 
 import YoutubeSearchResult from "./YoutubeSearchResult";
